% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndophenotypeExplorer.R
\name{EndophenotypeExplorer-class}
\alias{EndophenotypeExplorer-class}
\alias{EndophenotypeExplorer}
\title{EndophenotypeExplorer}
\value{
the url
}
\description{
R6 Class for exploring associations between eQTLs, variants, gene expression and gene regulation

An endophenotypeExplore has access to eQTLs, variant calls, and more
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EndophenotypeExplorer-new}{\code{EndophenotypeExplorer$new()}}
\item \href{#method-EndophenotypeExplorer-setTargetGene}{\code{EndophenotypeExplorer$setTargetGene()}}
\item \href{#method-EndophenotypeExplorer-setTargetChromosome}{\code{EndophenotypeExplorer$setTargetChromosome()}}
\item \href{#method-EndophenotypeExplorer-get.rna.matrix.codes}{\code{EndophenotypeExplorer$get.rna.matrix.codes()}}
\item \href{#method-EndophenotypeExplorer-get.rna.matrix}{\code{EndophenotypeExplorer$get.rna.matrix()}}
\item \href{#method-EndophenotypeExplorer-setupVcfURL}{\code{EndophenotypeExplorer$setupVcfURL()}}
\item \href{#method-EndophenotypeExplorer-getVcfUrl}{\code{EndophenotypeExplorer$getVcfUrl()}}
\item \href{#method-EndophenotypeExplorer-getGenoMatrix}{\code{EndophenotypeExplorer$getGenoMatrix()}}
\item \href{#method-EndophenotypeExplorer-getGenoMatrixByRSID}{\code{EndophenotypeExplorer$getGenoMatrixByRSID()}}
\item \href{#method-EndophenotypeExplorer-get.ampad.EQTLsForGene}{\code{EndophenotypeExplorer$get.ampad.EQTLsForGene()}}
\item \href{#method-EndophenotypeExplorer-getEQTLsInRegion}{\code{EndophenotypeExplorer$getEQTLsInRegion()}}
\item \href{#method-EndophenotypeExplorer-getIdMap}{\code{EndophenotypeExplorer$getIdMap()}}
\item \href{#method-EndophenotypeExplorer-biomart.rsidToLoc}{\code{EndophenotypeExplorer$biomart.rsidToLoc()}}
\item \href{#method-EndophenotypeExplorer-rsidToLoc}{\code{EndophenotypeExplorer$rsidToLoc()}}
\item \href{#method-EndophenotypeExplorer-new.rsidToLoc}{\code{EndophenotypeExplorer$new.rsidToLoc()}}
\item \href{#method-EndophenotypeExplorer-biomart.locsToRSID}{\code{EndophenotypeExplorer$biomart.locsToRSID()}}
\item \href{#method-EndophenotypeExplorer-locsToRSID}{\code{EndophenotypeExplorer$locsToRSID()}}
\item \href{#method-EndophenotypeExplorer-mapSampleIdToPatientAndCohort}{\code{EndophenotypeExplorer$mapSampleIdToPatientAndCohort()}}
\item \href{#method-EndophenotypeExplorer-setupClinicalData}{\code{EndophenotypeExplorer$setupClinicalData()}}
\item \href{#method-EndophenotypeExplorer-getClinicalTable}{\code{EndophenotypeExplorer$getClinicalTable()}}
\item \href{#method-EndophenotypeExplorer-setupGWASData}{\code{EndophenotypeExplorer$setupGWASData()}}
\item \href{#method-EndophenotypeExplorer-getGWASTables}{\code{EndophenotypeExplorer$getGWASTables()}}
\item \href{#method-EndophenotypeExplorer-identifyTargetGeneChromosome}{\code{EndophenotypeExplorer$identifyTargetGeneChromosome()}}
\item \href{#method-EndophenotypeExplorer-get.rosmap.patient.table}{\code{EndophenotypeExplorer$get.rosmap.patient.table()}}
\item \href{#method-EndophenotypeExplorer-get.sinai.patient.table}{\code{EndophenotypeExplorer$get.sinai.patient.table()}}
\item \href{#method-EndophenotypeExplorer-get.mayo.patient.table}{\code{EndophenotypeExplorer$get.mayo.patient.table()}}
\item \href{#method-EndophenotypeExplorer-sampleID.to.clinicalTable}{\code{EndophenotypeExplorer$sampleID.to.clinicalTable()}}
\item \href{#method-EndophenotypeExplorer-getStandardClinicalColumnNames}{\code{EndophenotypeExplorer$getStandardClinicalColumnNames()}}
\item \href{#method-EndophenotypeExplorer-standardizeMayoPatientTable}{\code{EndophenotypeExplorer$standardizeMayoPatientTable()}}
\item \href{#method-EndophenotypeExplorer-standardizeRosmapPatientTable}{\code{EndophenotypeExplorer$standardizeRosmapPatientTable()}}
\item \href{#method-EndophenotypeExplorer-standardizeSinaiPatientTable}{\code{EndophenotypeExplorer$standardizeSinaiPatientTable()}}
\item \href{#method-EndophenotypeExplorer-getAggregatedAlleleFrequencies}{\code{EndophenotypeExplorer$getAggregatedAlleleFrequencies()}}
\item \href{#method-EndophenotypeExplorer-getTissueExpression}{\code{EndophenotypeExplorer$getTissueExpression()}}
\item \href{#method-EndophenotypeExplorer-splitExpressionMatrixByMutationStatusAtRSID}{\code{EndophenotypeExplorer$splitExpressionMatrixByMutationStatusAtRSID()}}
\item \href{#method-EndophenotypeExplorer-splitMayoRnaMatrixByGenotype}{\code{EndophenotypeExplorer$splitMayoRnaMatrixByGenotype()}}
\item \href{#method-EndophenotypeExplorer-splitSinaiRnaMatrixByGenotype}{\code{EndophenotypeExplorer$splitSinaiRnaMatrixByGenotype()}}
\item \href{#method-EndophenotypeExplorer-splitRosmapRnaMatrixByGenotype}{\code{EndophenotypeExplorer$splitRosmapRnaMatrixByGenotype()}}
\item \href{#method-EndophenotypeExplorer-trenaScoreGenotypeStratifiedExpression}{\code{EndophenotypeExplorer$trenaScoreGenotypeStratifiedExpression()}}
\item \href{#method-EndophenotypeExplorer-subsetAndRelabelGenoMatrixByPatientIDs}{\code{EndophenotypeExplorer$subsetAndRelabelGenoMatrixByPatientIDs()}}
\item \href{#method-EndophenotypeExplorer-extremeBraakScoreSignificanceAtRSID}{\code{EndophenotypeExplorer$extremeBraakScoreSignificanceAtRSID()}}
\item \href{#method-EndophenotypeExplorer-clone}{\code{EndophenotypeExplorer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-new"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-new}{}}}
\subsection{Method \code{new()}}{
Create a new EndophenotypeExplorer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$new(
  target.gene,
  default.genome,
  vcf.project,
  chromosome = NA,
  verbose = FALSE,
  initialize.snpLocs = FALSE,
  defer.setupClinicalData.toSupportTesting = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target.gene}}{Gene of interest.}

\item{\code{default.genome}}{UCSC code, either `hg19` or `hg38`.}

\item{\code{vcf.project}}{character, either "ADNI" or "AMPAD"}

\item{\code{chromosome}}{default NA, may be unusual cases where we supply it (C2orf90/ECRG4) for instance}

\item{\code{verbose}}{logical}

\item{\code{initialize.snpLocs}}{logical, force this ~60 second process at startup}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `EndophenotypeExplorer` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-setTargetGene"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-setTargetGene}{}}}
\subsection{Method \code{setTargetGene()}}{
the targetGene and genome are widely referenced - this member function allows you
to change your initial choices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$setTargetGene(targetGene, genome, chromosome = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genome}}{UCSC code, either `hg19` or `hg38`.}

\item{\code{chromosome}}{default NA, causing a lookup}

\item{\code{target.gene}}{Gene of interest.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-setTargetChromosome"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-setTargetChromosome}{}}}
\subsection{Method \code{setTargetChromosome()}}{
the target chromosome and genome are widely referenced - this member function allows you
to change your initial choices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$setTargetChromosome(target.chromosome, genome)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genome}}{UCSC code, either `hg19` or `hg38`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-get.rna.matrix.codes"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-get.rna.matrix.codes}{}}}
\subsection{Method \code{get.rna.matrix.codes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$get.rna.matrix.codes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-get.rna.matrix"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-get.rna.matrix}{}}}
\subsection{Method \code{get.rna.matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$get.rna.matrix(code)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-setupVcfURL"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-setupVcfURL}{}}}
\subsection{Method \code{setupVcfURL()}}{
our genotypes are stored per-chromosome in vcf files.  this
function uses the current chromosome of interest, to provide
otherwise simple access to genotypes of the current gene and its
putative regulatory regions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$setupVcfURL(chromosome)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-getVcfUrl"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-getVcfUrl}{}}}
\subsection{Method \code{getVcfUrl()}}{
easy access to the current value
returns url in character form
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getVcfUrl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-getGenoMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-getGenoMatrix}{}}}
\subsection{Method \code{getGenoMatrix()}}{
genotypes for all samples in the specified locus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getGenoMatrix(chrom, start, end)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chrom}}{character}

\item{\code{start}}{numeric}

\item{\code{end}}{numeric
returns matrix, location by sample}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-getGenoMatrixByRSID"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-getGenoMatrixByRSID}{}}}
\subsection{Method \code{getGenoMatrixByRSID()}}{
genotypes for all samples at the specified rsids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getGenoMatrixByRSID(rsids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rsids}}{character
returns matrix, location by sample}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-get.ampad.EQTLsForGene"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-get.ampad.EQTLsForGene}{}}}
\subsection{Method \code{get.ampad.EQTLsForGene()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$get.ampad.EQTLsForGene()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-getEQTLsInRegion"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-getEQTLsInRegion}{}}}
\subsection{Method \code{getEQTLsInRegion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getEQTLsInRegion(chrom, start, end)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-getIdMap"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-getIdMap}{}}}
\subsection{Method \code{getIdMap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getIdMap()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-biomart.rsidToLoc"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-biomart.rsidToLoc}{}}}
\subsection{Method \code{biomart.rsidToLoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$biomart.rsidToLoc(rsids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-rsidToLoc"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-rsidToLoc}{}}}
\subsection{Method \code{rsidToLoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$rsidToLoc(rsids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-new.rsidToLoc"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-new.rsidToLoc}{}}}
\subsection{Method \code{new.rsidToLoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$new.rsidToLoc(rsids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-biomart.locsToRSID"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-biomart.locsToRSID}{}}}
\subsection{Method \code{biomart.locsToRSID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$biomart.locsToRSID(locs, genome)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-locsToRSID"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-locsToRSID}{}}}
\subsection{Method \code{locsToRSID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$locsToRSID(locs, genome)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-mapSampleIdToPatientAndCohort"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-mapSampleIdToPatientAndCohort}{}}}
\subsection{Method \code{mapSampleIdToPatientAndCohort()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$mapSampleIdToPatientAndCohort(sampleID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-setupClinicalData"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-setupClinicalData}{}}}
\subsection{Method \code{setupClinicalData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$setupClinicalData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-getClinicalTable"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-getClinicalTable}{}}}
\subsection{Method \code{getClinicalTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getClinicalTable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-setupGWASData"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-setupGWASData}{}}}
\subsection{Method \code{setupGWASData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$setupGWASData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-getGWASTables"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-getGWASTables}{}}}
\subsection{Method \code{getGWASTables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getGWASTables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-identifyTargetGeneChromosome"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-identifyTargetGeneChromosome}{}}}
\subsection{Method \code{identifyTargetGeneChromosome()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$identifyTargetGeneChromosome(target.gene)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-get.rosmap.patient.table"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-get.rosmap.patient.table}{}}}
\subsection{Method \code{get.rosmap.patient.table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$get.rosmap.patient.table(patientID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-get.sinai.patient.table"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-get.sinai.patient.table}{}}}
\subsection{Method \code{get.sinai.patient.table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$get.sinai.patient.table(patientID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-get.mayo.patient.table"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-get.mayo.patient.table}{}}}
\subsection{Method \code{get.mayo.patient.table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$get.mayo.patient.table(patientID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-sampleID.to.clinicalTable"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-sampleID.to.clinicalTable}{}}}
\subsection{Method \code{sampleID.to.clinicalTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$sampleID.to.clinicalTable(sampleID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-getStandardClinicalColumnNames"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-getStandardClinicalColumnNames}{}}}
\subsection{Method \code{getStandardClinicalColumnNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getStandardClinicalColumnNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-standardizeMayoPatientTable"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-standardizeMayoPatientTable}{}}}
\subsection{Method \code{standardizeMayoPatientTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$standardizeMayoPatientTable(tbl)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-standardizeRosmapPatientTable"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-standardizeRosmapPatientTable}{}}}
\subsection{Method \code{standardizeRosmapPatientTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$standardizeRosmapPatientTable(tbl)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-standardizeSinaiPatientTable"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-standardizeSinaiPatientTable}{}}}
\subsection{Method \code{standardizeSinaiPatientTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$standardizeSinaiPatientTable(tbl)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-getAggregatedAlleleFrequencies"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-getAggregatedAlleleFrequencies}{}}}
\subsection{Method \code{getAggregatedAlleleFrequencies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getAggregatedAlleleFrequencies(rsid, quiet = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-getTissueExpression"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-getTissueExpression}{}}}
\subsection{Method \code{getTissueExpression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getTissueExpression(geneSymbol, tissue)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-splitExpressionMatrixByMutationStatusAtRSID"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-splitExpressionMatrixByMutationStatusAtRSID}{}}}
\subsection{Method \code{splitExpressionMatrixByMutationStatusAtRSID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$splitExpressionMatrixByMutationStatusAtRSID(
  mtx,
  rsid,
  study.name
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-splitMayoRnaMatrixByGenotype"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-splitMayoRnaMatrixByGenotype}{}}}
\subsection{Method \code{splitMayoRnaMatrixByGenotype()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$splitMayoRnaMatrixByGenotype(mtx, rsid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-splitSinaiRnaMatrixByGenotype"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-splitSinaiRnaMatrixByGenotype}{}}}
\subsection{Method \code{splitSinaiRnaMatrixByGenotype()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$splitSinaiRnaMatrixByGenotype(mtx, rsid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-splitRosmapRnaMatrixByGenotype"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-splitRosmapRnaMatrixByGenotype}{}}}
\subsection{Method \code{splitRosmapRnaMatrixByGenotype()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$splitRosmapRnaMatrixByGenotype(mtx, rsid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-trenaScoreGenotypeStratifiedExpression"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-trenaScoreGenotypeStratifiedExpression}{}}}
\subsection{Method \code{trenaScoreGenotypeStratifiedExpression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$trenaScoreGenotypeStratifiedExpression(
  mtx.1,
  mtx.2,
  targetGene,
  tfs
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-subsetAndRelabelGenoMatrixByPatientIDs"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-subsetAndRelabelGenoMatrixByPatientIDs}{}}}
\subsection{Method \code{subsetAndRelabelGenoMatrixByPatientIDs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$subsetAndRelabelGenoMatrixByPatientIDs(
  mtx.geno,
  studyName
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-extremeBraakScoreSignificanceAtRSID"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-extremeBraakScoreSignificanceAtRSID}{}}}
\subsection{Method \code{extremeBraakScoreSignificanceAtRSID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$extremeBraakScoreSignificanceAtRSID(rsid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EndophenotypeExplorer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EndophenotypeExplorer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
