% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndophenotypeExplorer.R
\name{EndophenotypeExplorer-class}
\alias{EndophenotypeExplorer-class}
\alias{EndophenotypeExplorer}
\title{EndophenotypeExplorer}
\value{
the url
}
\description{
R6 Class for exploring associations between eQTLs, variants, gene expression and gene regulation

An endophenotypeExplore has access to eQTLs, variant calls, and more
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EndophenotypeExplorer$new()}}
\item \href{#method-setTargetGene}{\code{EndophenotypeExplorer$setTargetGene()}}
\item \href{#method-get.rna.matrix.codes}{\code{EndophenotypeExplorer$get.rna.matrix.codes()}}
\item \href{#method-get.rna.matrix}{\code{EndophenotypeExplorer$get.rna.matrix()}}
\item \href{#method-setupVcfURL}{\code{EndophenotypeExplorer$setupVcfURL()}}
\item \href{#method-getVcfUrl}{\code{EndophenotypeExplorer$getVcfUrl()}}
\item \href{#method-getGenoMatrix}{\code{EndophenotypeExplorer$getGenoMatrix()}}
\item \href{#method-getGenoMatrixByRSID}{\code{EndophenotypeExplorer$getGenoMatrixByRSID()}}
\item \href{#method-getEQTLsForGene}{\code{EndophenotypeExplorer$getEQTLsForGene()}}
\item \href{#method-getEQTLsInRegion}{\code{EndophenotypeExplorer$getEQTLsInRegion()}}
\item \href{#method-getIdMap}{\code{EndophenotypeExplorer$getIdMap()}}
\item \href{#method-biomart.rsidToLoc}{\code{EndophenotypeExplorer$biomart.rsidToLoc()}}
\item \href{#method-rsidToLoc}{\code{EndophenotypeExplorer$rsidToLoc()}}
\item \href{#method-biomart.locsToRSID}{\code{EndophenotypeExplorer$biomart.locsToRSID()}}
\item \href{#method-locsToRSID}{\code{EndophenotypeExplorer$locsToRSID()}}
\item \href{#method-mapSampleIdToPatientAndCohort}{\code{EndophenotypeExplorer$mapSampleIdToPatientAndCohort()}}
\item \href{#method-setupClinicalData}{\code{EndophenotypeExplorer$setupClinicalData()}}
\item \href{#method-getClinicalTable}{\code{EndophenotypeExplorer$getClinicalTable()}}
\item \href{#method-setupGWASData}{\code{EndophenotypeExplorer$setupGWASData()}}
\item \href{#method-getGWASTables}{\code{EndophenotypeExplorer$getGWASTables()}}
\item \href{#method-identifyTargetGeneChromosome}{\code{EndophenotypeExplorer$identifyTargetGeneChromosome()}}
\item \href{#method-get.rosmap.patient.table}{\code{EndophenotypeExplorer$get.rosmap.patient.table()}}
\item \href{#method-get.sinai.patient.table}{\code{EndophenotypeExplorer$get.sinai.patient.table()}}
\item \href{#method-get.mayo.patient.table}{\code{EndophenotypeExplorer$get.mayo.patient.table()}}
\item \href{#method-sampleID.to.clinicalTable}{\code{EndophenotypeExplorer$sampleID.to.clinicalTable()}}
\item \href{#method-getStandardClinicalColumnNames}{\code{EndophenotypeExplorer$getStandardClinicalColumnNames()}}
\item \href{#method-standardizeMayoPatientTable}{\code{EndophenotypeExplorer$standardizeMayoPatientTable()}}
\item \href{#method-standardizeRosmapPatientTable}{\code{EndophenotypeExplorer$standardizeRosmapPatientTable()}}
\item \href{#method-standardizeSinaiPatientTable}{\code{EndophenotypeExplorer$standardizeSinaiPatientTable()}}
\item \href{#method-getAggregatedAlleleFrequencies}{\code{EndophenotypeExplorer$getAggregatedAlleleFrequencies()}}
\item \href{#method-getTissueExpression}{\code{EndophenotypeExplorer$getTissueExpression()}}
\item \href{#method-splitExpressionMatrixByMutationStatusAtRSID}{\code{EndophenotypeExplorer$splitExpressionMatrixByMutationStatusAtRSID()}}
\item \href{#method-splitMayoRnaMatrixByGenotype}{\code{EndophenotypeExplorer$splitMayoRnaMatrixByGenotype()}}
\item \href{#method-splitSinaiRnaMatrixByGenotype}{\code{EndophenotypeExplorer$splitSinaiRnaMatrixByGenotype()}}
\item \href{#method-splitRosmapRnaMatrixByGenotype}{\code{EndophenotypeExplorer$splitRosmapRnaMatrixByGenotype()}}
\item \href{#method-trenaScoreGenotypeStratifiedExpression}{\code{EndophenotypeExplorer$trenaScoreGenotypeStratifiedExpression()}}
\item \href{#method-subsetAndRelabelGenoMatrixByPatientIDs}{\code{EndophenotypeExplorer$subsetAndRelabelGenoMatrixByPatientIDs()}}
\item \href{#method-clone}{\code{EndophenotypeExplorer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new EndophenotypeExplorer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$new(
  target.gene,
  default.genome,
  vcf.project = "AMPAD",
  verbose = FALSE,
  initialize.snpLocs = FALSE,
  defer.setupClinicalData.toSupportTesting = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target.gene}}{Gene of interest.}

\item{\code{default.genome}}{UCSC code, either `hg19` or `hg38`.}

\item{\code{vcf.project}}{character, either "ADNI" or "AMPAD", the latter by default}

\item{\code{verbose}}{logical}

\item{\code{initialize.snpLocs}}{logical, force this ~60 second process at startup}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `EndophenotypeExplorer` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setTargetGene"></a>}}
\if{latex}{\out{\hypertarget{method-setTargetGene}{}}}
\subsection{Method \code{setTargetGene()}}{
the targetGene and genome are widely referenced - this member function allows you
to change your initial choices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$setTargetGene(targetGene, genome)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genome}}{UCSC code, either `hg19` or `hg38`.}

\item{\code{target.gene}}{Gene of interest.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.rna.matrix.codes"></a>}}
\if{latex}{\out{\hypertarget{method-get.rna.matrix.codes}{}}}
\subsection{Method \code{get.rna.matrix.codes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$get.rna.matrix.codes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.rna.matrix"></a>}}
\if{latex}{\out{\hypertarget{method-get.rna.matrix}{}}}
\subsection{Method \code{get.rna.matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$get.rna.matrix(code)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupVcfURL"></a>}}
\if{latex}{\out{\hypertarget{method-setupVcfURL}{}}}
\subsection{Method \code{setupVcfURL()}}{
our genotypes are stored per-chromosome in vcf files.  this
function uses the current chromosome of interest, to provide
otherwise simple access to genotypes of the current gene and its
putative regulatory regions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$setupVcfURL(chromosome)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getVcfUrl"></a>}}
\if{latex}{\out{\hypertarget{method-getVcfUrl}{}}}
\subsection{Method \code{getVcfUrl()}}{
easy access to the current value
returns url in character form
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getVcfUrl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGenoMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-getGenoMatrix}{}}}
\subsection{Method \code{getGenoMatrix()}}{
genotypes for all samples in the specified locus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getGenoMatrix(chrom, start, end)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chrom}}{character}

\item{\code{start}}{numeric}

\item{\code{end}}{numeric
returns matrix, location by sample}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGenoMatrixByRSID"></a>}}
\if{latex}{\out{\hypertarget{method-getGenoMatrixByRSID}{}}}
\subsection{Method \code{getGenoMatrixByRSID()}}{
genotypes for all samples at the specified rsids
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getGenoMatrixByRSID(rsids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rsids}}{character
returns matrix, location by sample}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEQTLsForGene"></a>}}
\if{latex}{\out{\hypertarget{method-getEQTLsForGene}{}}}
\subsection{Method \code{getEQTLsForGene()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getEQTLsForGene()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEQTLsInRegion"></a>}}
\if{latex}{\out{\hypertarget{method-getEQTLsInRegion}{}}}
\subsection{Method \code{getEQTLsInRegion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getEQTLsInRegion(chrom, start, end)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIdMap"></a>}}
\if{latex}{\out{\hypertarget{method-getIdMap}{}}}
\subsection{Method \code{getIdMap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getIdMap()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biomart.rsidToLoc"></a>}}
\if{latex}{\out{\hypertarget{method-biomart.rsidToLoc}{}}}
\subsection{Method \code{biomart.rsidToLoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$biomart.rsidToLoc(rsids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rsidToLoc"></a>}}
\if{latex}{\out{\hypertarget{method-rsidToLoc}{}}}
\subsection{Method \code{rsidToLoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$rsidToLoc(rsids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biomart.locsToRSID"></a>}}
\if{latex}{\out{\hypertarget{method-biomart.locsToRSID}{}}}
\subsection{Method \code{biomart.locsToRSID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$biomart.locsToRSID(locs, genome)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-locsToRSID"></a>}}
\if{latex}{\out{\hypertarget{method-locsToRSID}{}}}
\subsection{Method \code{locsToRSID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$locsToRSID(locs, genome)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapSampleIdToPatientAndCohort"></a>}}
\if{latex}{\out{\hypertarget{method-mapSampleIdToPatientAndCohort}{}}}
\subsection{Method \code{mapSampleIdToPatientAndCohort()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$mapSampleIdToPatientAndCohort(sampleID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupClinicalData"></a>}}
\if{latex}{\out{\hypertarget{method-setupClinicalData}{}}}
\subsection{Method \code{setupClinicalData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$setupClinicalData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getClinicalTable"></a>}}
\if{latex}{\out{\hypertarget{method-getClinicalTable}{}}}
\subsection{Method \code{getClinicalTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getClinicalTable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupGWASData"></a>}}
\if{latex}{\out{\hypertarget{method-setupGWASData}{}}}
\subsection{Method \code{setupGWASData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$setupGWASData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGWASTables"></a>}}
\if{latex}{\out{\hypertarget{method-getGWASTables}{}}}
\subsection{Method \code{getGWASTables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getGWASTables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-identifyTargetGeneChromosome"></a>}}
\if{latex}{\out{\hypertarget{method-identifyTargetGeneChromosome}{}}}
\subsection{Method \code{identifyTargetGeneChromosome()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$identifyTargetGeneChromosome(target.gene)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.rosmap.patient.table"></a>}}
\if{latex}{\out{\hypertarget{method-get.rosmap.patient.table}{}}}
\subsection{Method \code{get.rosmap.patient.table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$get.rosmap.patient.table(patientID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.sinai.patient.table"></a>}}
\if{latex}{\out{\hypertarget{method-get.sinai.patient.table}{}}}
\subsection{Method \code{get.sinai.patient.table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$get.sinai.patient.table(patientID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.mayo.patient.table"></a>}}
\if{latex}{\out{\hypertarget{method-get.mayo.patient.table}{}}}
\subsection{Method \code{get.mayo.patient.table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$get.mayo.patient.table(patientID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sampleID.to.clinicalTable"></a>}}
\if{latex}{\out{\hypertarget{method-sampleID.to.clinicalTable}{}}}
\subsection{Method \code{sampleID.to.clinicalTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$sampleID.to.clinicalTable(sampleID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStandardClinicalColumnNames"></a>}}
\if{latex}{\out{\hypertarget{method-getStandardClinicalColumnNames}{}}}
\subsection{Method \code{getStandardClinicalColumnNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getStandardClinicalColumnNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standardizeMayoPatientTable"></a>}}
\if{latex}{\out{\hypertarget{method-standardizeMayoPatientTable}{}}}
\subsection{Method \code{standardizeMayoPatientTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$standardizeMayoPatientTable(tbl)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standardizeRosmapPatientTable"></a>}}
\if{latex}{\out{\hypertarget{method-standardizeRosmapPatientTable}{}}}
\subsection{Method \code{standardizeRosmapPatientTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$standardizeRosmapPatientTable(tbl)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standardizeSinaiPatientTable"></a>}}
\if{latex}{\out{\hypertarget{method-standardizeSinaiPatientTable}{}}}
\subsection{Method \code{standardizeSinaiPatientTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$standardizeSinaiPatientTable(tbl)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAggregatedAlleleFrequencies"></a>}}
\if{latex}{\out{\hypertarget{method-getAggregatedAlleleFrequencies}{}}}
\subsection{Method \code{getAggregatedAlleleFrequencies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getAggregatedAlleleFrequencies(rsid, quiet = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTissueExpression"></a>}}
\if{latex}{\out{\hypertarget{method-getTissueExpression}{}}}
\subsection{Method \code{getTissueExpression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getTissueExpression(geneSymbol, tissue)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-splitExpressionMatrixByMutationStatusAtRSID"></a>}}
\if{latex}{\out{\hypertarget{method-splitExpressionMatrixByMutationStatusAtRSID}{}}}
\subsection{Method \code{splitExpressionMatrixByMutationStatusAtRSID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$splitExpressionMatrixByMutationStatusAtRSID(
  mtx,
  rsid,
  study.name
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-splitMayoRnaMatrixByGenotype"></a>}}
\if{latex}{\out{\hypertarget{method-splitMayoRnaMatrixByGenotype}{}}}
\subsection{Method \code{splitMayoRnaMatrixByGenotype()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$splitMayoRnaMatrixByGenotype(mtx, rsid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-splitSinaiRnaMatrixByGenotype"></a>}}
\if{latex}{\out{\hypertarget{method-splitSinaiRnaMatrixByGenotype}{}}}
\subsection{Method \code{splitSinaiRnaMatrixByGenotype()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$splitSinaiRnaMatrixByGenotype(mtx, rsid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-splitRosmapRnaMatrixByGenotype"></a>}}
\if{latex}{\out{\hypertarget{method-splitRosmapRnaMatrixByGenotype}{}}}
\subsection{Method \code{splitRosmapRnaMatrixByGenotype()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$splitRosmapRnaMatrixByGenotype(mtx, rsid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trenaScoreGenotypeStratifiedExpression"></a>}}
\if{latex}{\out{\hypertarget{method-trenaScoreGenotypeStratifiedExpression}{}}}
\subsection{Method \code{trenaScoreGenotypeStratifiedExpression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$trenaScoreGenotypeStratifiedExpression(
  mtx.1,
  mtx.2,
  targetGene,
  tfs
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subsetAndRelabelGenoMatrixByPatientIDs"></a>}}
\if{latex}{\out{\hypertarget{method-subsetAndRelabelGenoMatrixByPatientIDs}{}}}
\subsection{Method \code{subsetAndRelabelGenoMatrixByPatientIDs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$subsetAndRelabelGenoMatrixByPatientIDs(
  mtx.geno,
  studyName
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
