% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndophenotypeExplorer.R
\name{EndophenotypeExplorer-class}
\alias{EndophenotypeExplorer-class}
\alias{EndophenotypeExplorer}
\alias{R6}
\alias{Class}
\alias{for}
\alias{exploring}
\alias{associations}
\alias{between}
\alias{eQTLs,}
\alias{variants,}
\alias{gene}
\alias{expression}
\alias{and}
\alias{regulation}
\alias{An}
\alias{endophenotypeExplore}
\alias{has}
\alias{access}
\alias{to}
\alias{variant}
\alias{calls,}
\alias{more}
\title{EndophenotypeExplorer}
\description{
EndophenotypeExplorer

EndophenotypeExplorer
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EndophenotypeExplorer$new()}}
\item \href{#method-setupClinicalData}{\code{EndophenotypeExplorer$setupClinicalData()}}
\item \href{#method-vcfSampleID.to.clinicalTable}{\code{EndophenotypeExplorer$vcfSampleID.to.clinicalTable()}}
\item \href{#method-clone}{\code{EndophenotypeExplorer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new EndophenotypeExplorer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$new(target.gene, default.genome, vcf.url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target.gene}}{Gene of interest.}

\item{\code{default.genome}}{UCSC code, either `hg19` or `hg38`.}

\item{\code{vcf.url}}{https endpoint from serving indexed vcf files}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `EndophenotypeExplorer` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupClinicalData"></a>}}
\if{latex}{\out{\hypertarget{method-setupClinicalData}{}}}
\subsection{Method \code{setupClinicalData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$setupClinicalData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vcfSampleID.to.clinicalTable"></a>}}
\if{latex}{\out{\hypertarget{method-vcfSampleID.to.clinicalTable}{}}}
\subsection{Method \code{vcfSampleID.to.clinicalTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$vcfSampleID.to.clinicalTable(sampleIDs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
