% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndophenotypeExplorer.R
\name{EndophenotypeExplorer-class}
\alias{EndophenotypeExplorer-class}
\alias{EndophenotypeExplorer}
\alias{R6}
\alias{Class}
\alias{for}
\alias{exploring}
\alias{associations}
\alias{between}
\alias{eQTLs,}
\alias{variants,}
\alias{gene}
\alias{expression}
\alias{and}
\alias{regulation}
\alias{An}
\alias{endophenotypeExplore}
\alias{has}
\alias{access}
\alias{to}
\alias{variant}
\alias{calls,}
\alias{more}
\title{EndophenotypeExplorer}
\description{
EndophenotypeExplorer

EndophenotypeExplorer
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EndophenotypeExplorer$new()}}
\item \href{#method-setupVcfURL}{\code{EndophenotypeExplorer$setupVcfURL()}}
\item \href{#method-getVcfUrl}{\code{EndophenotypeExplorer$getVcfUrl()}}
\item \href{#method-getGenoMatrix}{\code{EndophenotypeExplorer$getGenoMatrix()}}
\item \href{#method-getGenoMatrixByRSID}{\code{EndophenotypeExplorer$getGenoMatrixByRSID()}}
\item \href{#method-getEQTLsForGene}{\code{EndophenotypeExplorer$getEQTLsForGene()}}
\item \href{#method-getIdMap}{\code{EndophenotypeExplorer$getIdMap()}}
\item \href{#method-rsidToLoc}{\code{EndophenotypeExplorer$rsidToLoc()}}
\item \href{#method-locsToRSID}{\code{EndophenotypeExplorer$locsToRSID()}}
\item \href{#method-mapSampleIdToPatientAndCohort}{\code{EndophenotypeExplorer$mapSampleIdToPatientAndCohort()}}
\item \href{#method-setupClinicalData}{\code{EndophenotypeExplorer$setupClinicalData()}}
\item \href{#method-getClinicalTable}{\code{EndophenotypeExplorer$getClinicalTable()}}
\item \href{#method-setupGWASData}{\code{EndophenotypeExplorer$setupGWASData()}}
\item \href{#method-getGWASTables}{\code{EndophenotypeExplorer$getGWASTables()}}
\item \href{#method-identifyTargetGeneChromosome}{\code{EndophenotypeExplorer$identifyTargetGeneChromosome()}}
\item \href{#method-get.rosmap.patient.table}{\code{EndophenotypeExplorer$get.rosmap.patient.table()}}
\item \href{#method-get.sinai.patient.table}{\code{EndophenotypeExplorer$get.sinai.patient.table()}}
\item \href{#method-get.mayo.patient.table}{\code{EndophenotypeExplorer$get.mayo.patient.table()}}
\item \href{#method-sampleID.to.clinicalTable}{\code{EndophenotypeExplorer$sampleID.to.clinicalTable()}}
\item \href{#method-getStandardClinicalColumnNames}{\code{EndophenotypeExplorer$getStandardClinicalColumnNames()}}
\item \href{#method-standardizeMayoPatientTable}{\code{EndophenotypeExplorer$standardizeMayoPatientTable()}}
\item \href{#method-standardizeRosmapPatientTable}{\code{EndophenotypeExplorer$standardizeRosmapPatientTable()}}
\item \href{#method-standardizeSinaiPatientTable}{\code{EndophenotypeExplorer$standardizeSinaiPatientTable()}}
\item \href{#method-getAggregatedAlleleFrequencies}{\code{EndophenotypeExplorer$getAggregatedAlleleFrequencies()}}
\item \href{#method-getTissueExpression}{\code{EndophenotypeExplorer$getTissueExpression()}}
\item \href{#method-clone}{\code{EndophenotypeExplorer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new EndophenotypeExplorer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$new(
  target.gene,
  default.genome,
  verbose = FALSE,
  defer.setupClinicalData.toSupportTesting = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target.gene}}{Gene of interest.}

\item{\code{default.genome}}{UCSC code, either `hg19` or `hg38`.}

\item{\code{vcf.url}}{https endpoint from serving indexed vcf files}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `EndophenotypeExplorer` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupVcfURL"></a>}}
\if{latex}{\out{\hypertarget{method-setupVcfURL}{}}}
\subsection{Method \code{setupVcfURL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$setupVcfURL(chromosome)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getVcfUrl"></a>}}
\if{latex}{\out{\hypertarget{method-getVcfUrl}{}}}
\subsection{Method \code{getVcfUrl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getVcfUrl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGenoMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-getGenoMatrix}{}}}
\subsection{Method \code{getGenoMatrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getGenoMatrix(chrom, start, end)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGenoMatrixByRSID"></a>}}
\if{latex}{\out{\hypertarget{method-getGenoMatrixByRSID}{}}}
\subsection{Method \code{getGenoMatrixByRSID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getGenoMatrixByRSID(rsids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEQTLsForGene"></a>}}
\if{latex}{\out{\hypertarget{method-getEQTLsForGene}{}}}
\subsection{Method \code{getEQTLsForGene()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getEQTLsForGene()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIdMap"></a>}}
\if{latex}{\out{\hypertarget{method-getIdMap}{}}}
\subsection{Method \code{getIdMap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getIdMap()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rsidToLoc"></a>}}
\if{latex}{\out{\hypertarget{method-rsidToLoc}{}}}
\subsection{Method \code{rsidToLoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$rsidToLoc(rsids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-locsToRSID"></a>}}
\if{latex}{\out{\hypertarget{method-locsToRSID}{}}}
\subsection{Method \code{locsToRSID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$locsToRSID(locs, genome)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapSampleIdToPatientAndCohort"></a>}}
\if{latex}{\out{\hypertarget{method-mapSampleIdToPatientAndCohort}{}}}
\subsection{Method \code{mapSampleIdToPatientAndCohort()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$mapSampleIdToPatientAndCohort(sampleID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupClinicalData"></a>}}
\if{latex}{\out{\hypertarget{method-setupClinicalData}{}}}
\subsection{Method \code{setupClinicalData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$setupClinicalData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getClinicalTable"></a>}}
\if{latex}{\out{\hypertarget{method-getClinicalTable}{}}}
\subsection{Method \code{getClinicalTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getClinicalTable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupGWASData"></a>}}
\if{latex}{\out{\hypertarget{method-setupGWASData}{}}}
\subsection{Method \code{setupGWASData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$setupGWASData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGWASTables"></a>}}
\if{latex}{\out{\hypertarget{method-getGWASTables}{}}}
\subsection{Method \code{getGWASTables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getGWASTables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-identifyTargetGeneChromosome"></a>}}
\if{latex}{\out{\hypertarget{method-identifyTargetGeneChromosome}{}}}
\subsection{Method \code{identifyTargetGeneChromosome()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$identifyTargetGeneChromosome(target.gene)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.rosmap.patient.table"></a>}}
\if{latex}{\out{\hypertarget{method-get.rosmap.patient.table}{}}}
\subsection{Method \code{get.rosmap.patient.table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$get.rosmap.patient.table(patientID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.sinai.patient.table"></a>}}
\if{latex}{\out{\hypertarget{method-get.sinai.patient.table}{}}}
\subsection{Method \code{get.sinai.patient.table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$get.sinai.patient.table(patientID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.mayo.patient.table"></a>}}
\if{latex}{\out{\hypertarget{method-get.mayo.patient.table}{}}}
\subsection{Method \code{get.mayo.patient.table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$get.mayo.patient.table(patientID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sampleID.to.clinicalTable"></a>}}
\if{latex}{\out{\hypertarget{method-sampleID.to.clinicalTable}{}}}
\subsection{Method \code{sampleID.to.clinicalTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$sampleID.to.clinicalTable(sampleID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStandardClinicalColumnNames"></a>}}
\if{latex}{\out{\hypertarget{method-getStandardClinicalColumnNames}{}}}
\subsection{Method \code{getStandardClinicalColumnNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getStandardClinicalColumnNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standardizeMayoPatientTable"></a>}}
\if{latex}{\out{\hypertarget{method-standardizeMayoPatientTable}{}}}
\subsection{Method \code{standardizeMayoPatientTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$standardizeMayoPatientTable(tbl)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standardizeRosmapPatientTable"></a>}}
\if{latex}{\out{\hypertarget{method-standardizeRosmapPatientTable}{}}}
\subsection{Method \code{standardizeRosmapPatientTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$standardizeRosmapPatientTable(tbl)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standardizeSinaiPatientTable"></a>}}
\if{latex}{\out{\hypertarget{method-standardizeSinaiPatientTable}{}}}
\subsection{Method \code{standardizeSinaiPatientTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$standardizeSinaiPatientTable(tbl)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAggregatedAlleleFrequencies"></a>}}
\if{latex}{\out{\hypertarget{method-getAggregatedAlleleFrequencies}{}}}
\subsection{Method \code{getAggregatedAlleleFrequencies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getAggregatedAlleleFrequencies(rsid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTissueExpression"></a>}}
\if{latex}{\out{\hypertarget{method-getTissueExpression}{}}}
\subsection{Method \code{getTissueExpression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$getTissueExpression(geneSymbol, tissue)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EndophenotypeExplorer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
